Wall Generation System is Working
While all wall tiles are not currently implemented, the system is rule based, where a number from -1 to 3 (inclusive) decides what tiles need to be surrounding a WallBase position for a WallSegment to be placed there.

Notes:
- You can no longer create walls without corners filled.
- You can no longer generate walls on the inner edge of the floor silhouette.
+ You can define a number of WallSegments for a WallDefinition instead of the constant 8 from before.

As for defining a WallSegment's required surrounding tile types:
-1 = Anything
0 = Air
1 = Floor
2 = Wall (another WallBase)
3 + Wall + Floor (Likely unused)

The system is currently unoptimized and increases Floor generation times by around 1500-2000ms. Future commits should focus on making the system work for all edge cases, however.

- Des