Refined Floor Drawing
Floor Drawing has changed. A good bit of the code has been cleaned up, and it runs a little faster. So far, it looks like the average full Floor generation time is around 3500ms to 3700ms.

The important changes to organization is as follows:
/ The Floor scene now expects the TileMap child to have its layers organized in a certain way. Layer 0 is for Hall Floors, 1 for Room Floors, 2 for Room Props (likely to be y-sorted in the future), 3 for Room Walls, and 4 for Floor Walls.
/ In turn, Floor.cs also expects the TileMaps of RoomDefinitions to have its layers organized similarly. Layer 0 is for Floors, 1 for Props (again, likely to be y-sorted in the future), and 2 for Walls.
- This all means that there cannot be more than 5 layers to the Floor TileMap, and no more than 3 for the Room TileMap. Layers are not created as needed during Floor.DrawFloor(). Only when Floor.Setup() is called.

Remember this when creating new RoomDefinitions. Good luck.

- Des