Made the Game Feel Like a Game
There is a lot of changes in this commit, so it should be considered a major one. I'm using a FSM + Behavior Tree workflow thanks to BehaviourToolkit. The game's logic has been refined drastically thanks to a main FSM that decides the overarching state of the game.

In a nutshell, here is what has changed (in no particular order):

+ Title screen was added - it is shown on game startup and you can start or quit the game
+ Loading screen was added - shown when the game is actively generating a new floor (as that takes a bit of time)
+ Pause menu was added - shown during the MainState by pressing esc or the "menu" button. You can return to the title screen by pressing q or the "quit" button. That hint is also included in the pause menu itself.
+ Placeholder states and transititons, basically outlining what GameManager should be managing.
/ When you start the game, you are no longer placed at (0, 0), but in the center of the starting room, which is the room furthest from (0, 0).
/ Game scenes like the player are now in the Pauseable process group which only allows their logic to run while get_tree().paused = false.
/ The "world" scene is deprecated.
/ The "game" scene is now the main scene for the project.
/ The "game" scene now only starts with GameManager as its only child node.

That should be it. To reiterate, this commit is large so consider it major. Remember that GameManager should only be responsible for the GameStateMachine. Good luck.

- Des